+++ examples/alice_http_client/lib/main.dart
@@ -22,7 +22,8 @@ class _MyAppState extends State<MyApp> {
   late final AliceHttpClientAdapter _httpClientAdapter =
       AliceHttpClientAdapter();
 
-  late final Alice _alice = Alice(configuration: AliceConfiguration())
+  final configuration = AliceConfiguration();
+  late final Alice _alice = Alice(configuration: configuration)
     ..addAdapter(_httpClientAdapter);
 
   @override