+++ packages/alice/lib/ui/calls_list/widget/alice_logs_screen.dart
@@ -23,12 +23,10 @@ class AliceLogsScreen extends StatelessWidget {
           ? AliceRawLogListWidget(
               scrollController: scrollController,
               getRawLogs: aliceLogger?.getAndroidRawLogs(),
-              emptyWidget: const AliceEmptyLogsWidget(),
             )
           : AliceLogListWidget(
               logsStream: aliceLogger?.logsStream,
               scrollController: scrollController,
-              emptyWidget: const AliceEmptyLogsWidget(),
             )
       : const AliceEmptyLogsWidget();
 }