+++ packages/alice/lib/ui/calls_list/widget/alice_raw_log_list_widger.dart
@@ -1,4 +1,6 @@
 import 'package:alice/model/alice_translation.dart';
+import 'package:alice/ui/calls_list/widget/alice_empty_logs_widget.dart';
+import 'package:alice/ui/calls_list/widget/alice_error_logs_widget.dart';
 import 'package:alice/ui/common/alice_context_ext.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -8,45 +10,53 @@ class AliceRawLogListWidget extends StatelessWidget {
   const AliceRawLogListWidget({
     required this.scrollController,
     required this.getRawLogs,
-    required this.emptyWidget,
     super.key,
   });
 
   final ScrollController scrollController;
   final Future<String>? getRawLogs;
-  final Widget emptyWidget;
 
   @override
   Widget build(BuildContext context) {
     return FutureBuilder<String>(
       future: getRawLogs,
       builder: (context, AsyncSnapshot<String> snapshot) {
-        if (snapshot.hasData) {
-          if (snapshot.data?.isNotEmpty == true) {
-            return Scrollbar(
-              thickness: 8,
-              controller: scrollController,
-              child: SingleChildScrollView(
-                controller: scrollController,
-                child: Padding(
-                  padding: const EdgeInsets.all(8),
-                  child: InkWell(
-                    onLongPress: () => _copyToClipboard(
-                      context: context,
-                      text: snapshot.data ?? '',
-                    ),
-                    child: Text(
-                      snapshot.data ?? '',
-                      style: const TextStyle(fontSize: 10),
-                    ),
-                  ),
+        if (snapshot.connectionState == ConnectionState.none ||
+            snapshot.connectionState == ConnectionState.waiting) {
+          return const Center(
+            child: CircularProgressIndicator(),
+          );
+        }
+
+        if (snapshot.hasError) {
+          return const AliceErrorLogsWidget();
+        }
+
+        final logs = snapshot.data ?? '';
+        if (logs.isEmpty) {
+          return const AliceEmptyLogsWidget();
+        }
+
+        return Scrollbar(
+          thickness: 8,
+          controller: scrollController,
+          child: SingleChildScrollView(
+            controller: scrollController,
+            child: Padding(
+              padding: const EdgeInsets.all(8),
+              child: InkWell(
+                onLongPress: () => _copyToClipboard(
+                  context: context,
+                  text: snapshot.data ?? '',
+                ),
+                child: Text(
+                  snapshot.data ?? '',
+                  style: const TextStyle(fontSize: 10),
                 ),
               ),
-            );
-          }
-          return emptyWidget;
-        }
-        return const Center(child: CircularProgressIndicator());
+            ),
+          ),
+        );
       },
     );
   }