+++ packages/alice/CHANGELOG.md
@@ -1,6 +1,6 @@
 # 1.0.0-dev.10
 
-* [BREAKING_CHANGE] Removed maxCallsCount. To change call count, use `storage` constructor parameter
+* [BREAKING_CHANGE] Removed `maxCallsCount`. To change call count, use `storage` constructor parameter
   with call count.
 * [BREAKING_CHANGE] Replaced configuration parameter of Alice with `AliceConfiguration`.
 * Fixed issue with invalid count of calls in notification.